+++ src/sagemaker_pytorch_serving_container/torchserve.py
@@ -24,13 +24,13 @@ import logging
 from retrying import retry
 
 import sagemaker_pytorch_serving_container
-from sagemaker_inference import default_handler_service, environment, utils
+from sagemaker_pytorch_serving_container import ts_environment
+from sagemaker_inference import environment, utils
 from sagemaker_inference.environment import code_dir
 
 logger = logging.getLogger()
 
 TS_CONFIG_FILE = os.path.join("/etc", "sagemaker-ts.properties")
-DEFAULT_HANDLER_SERVICE = default_handler_service.__name__
 DEFAULT_TS_CONFIG_FILE = pkg_resources.resource_filename(
     sagemaker_pytorch_serving_container.__name__, "/etc/default-ts.properties"
 )
@@ -38,15 +38,13 @@ MME_TS_CONFIG_FILE = pkg_resources.resource_filename(
     sagemaker_pytorch_serving_container.__name__, "/etc/mme-ts.properties"
 )
 DEFAULT_TS_LOG_FILE = pkg_resources.resource_filename(
-    sagemaker_pytorch_serving_container.__name__, "/etc/log4j.properties"
+    sagemaker_pytorch_serving_container.__name__, "/etc/log4j2.xml"
 )
-DEFAULT_TS_MODEL_DIRECTORY = os.path.join(os.getcwd(), ".sagemaker", "ts", "models")
 DEFAULT_TS_MODEL_NAME = "model"
-DEFAULT_TS_CODE_DIR = "code"
 DEFAULT_HANDLER_SERVICE = "sagemaker_pytorch_serving_container.handler_service"
 
 ENABLE_MULTI_MODEL = os.getenv("SAGEMAKER_MULTI_MODEL", "false") == "true"
-MODEL_STORE = "/" if ENABLE_MULTI_MODEL else DEFAULT_TS_MODEL_DIRECTORY
+MODEL_STORE = "/" if ENABLE_MULTI_MODEL else os.path.join(os.getcwd(), ".sagemaker", "ts", "models")
 
 PYTHON_PATH_ENV = "PYTHONPATH"
 REQUIREMENTS_PATH = os.path.join(code_dir, "requirements.txt")
@@ -72,11 +70,13 @@ def start_torchserve(handler_service=DEFAULT_HANDLER_SERVICE):
     if ENABLE_MULTI_MODEL:
         if "SAGEMAKER_HANDLER" not in os.environ:
             os.environ["SAGEMAKER_HANDLER"] = handler_service
-        _set_python_path()
     else:
-        _adapt_to_ts_format(handler_service)
+        if not os.path.exists(MODEL_STORE):
+            os.makedirs(MODEL_STORE)
 
-    _create_torchserve_config_file()
+    _set_python_path()
+
+    _create_torchserve_config_file(handler_service)
 
     if os.path.exists(REQUIREMENTS_PATH):
         _install_requirements()
@@ -91,7 +91,7 @@ def start_torchserve(handler_service=DEFAULT_HANDLER_SERVICE):
         "--log-config",
         DEFAULT_TS_LOG_FILE,
         "--models",
-        "model.mar"
+        DEFAULT_TS_MODEL_NAME + "=" + environment.model_dir
     ]
 
     print(ts_torchserve_cmd)
@@ -106,30 +106,6 @@ def start_torchserve(handler_service=DEFAULT_HANDLER_SERVICE):
     ts_process.wait()
 
 
-def _adapt_to_ts_format(handler_service):
-    if not os.path.exists(DEFAULT_TS_MODEL_DIRECTORY):
-        os.makedirs(DEFAULT_TS_MODEL_DIRECTORY)
-
-    model_archiver_cmd = [
-        "torch-model-archiver",
-        "--model-name",
-        DEFAULT_TS_MODEL_NAME,
-        "--handler",
-        handler_service,
-        "--export-path",
-        DEFAULT_TS_MODEL_DIRECTORY,
-        "--version",
-        "1",
-        "--extra-files",
-        os.path.join(environment.model_dir)
-    ]
-
-    logger.info(model_archiver_cmd)
-    subprocess.check_call(model_archiver_cmd)
-
-    _set_python_path()
-
-
 def _set_python_path():
     # Torchserve handles code execution by appending the export path, provided
     # to the model archiver, to the PYTHONPATH env var.
@@ -141,21 +117,41 @@ def _set_python_path():
         os.environ[PYTHON_PATH_ENV] = environment.code_dir
 
 
-def _create_torchserve_config_file():
-    configuration_properties = _generate_ts_config_properties()
+def _create_torchserve_config_file(handler_service):
+    configuration_properties = _generate_ts_config_properties(handler_service)
 
     utils.write_file(TS_CONFIG_FILE, configuration_properties)
 
 
-def _generate_ts_config_properties():
+def _generate_ts_config_properties(handler_service):
     env = environment.Environment()
-
     user_defined_configuration = {
         "vmargs":"-XX:-UseContainerSupport -XX:InitialRAMPercentage=8.0 -XX:MaxRAMPercentage=10.0 -XX:-UseLargePages -XX:+UseG1GC -XX:+ExitOnOutOfMemoryError",        "default_response_timeout": env.model_server_timeout,
         "default_workers_per_model": env.model_server_workers,
         "inference_address": "http://0.0.0.0:{}".format(env.inference_http_port),
         "management_address": "http://0.0.0.0:{}".format(env.management_http_port),
+        "default_service_handler": handler_service + ":handle",
     }
+
+    ts_env = ts_environment.TorchServeEnvironment()
+
+    if ts_env.is_env_set() and not ENABLE_MULTI_MODEL:
+        models_string = f'''{{\\
+        "{DEFAULT_TS_MODEL_NAME}": {{\\
+            "1.0": {{\\
+                "defaultVersion": true,\\
+                "marName": "{DEFAULT_TS_MODEL_NAME}.mar",\\
+                "minWorkers": {ts_env._min_workers},\\
+                "maxWorkers": {ts_env._max_workers},\\
+                "batchSize": {ts_env._batch_size},\\
+                "maxBatchDelay": {ts_env._max_batch_delay},\\
+                "responseTimeout": {ts_env._response_timeout}\\
+                }}\\
+            }}\\
+        }}'''
+        user_defined_configuration["models"] = models_string
+        logger.warn("Sagemaker TS environment variables have been set and will be used "
+                    "for single model endpoint.")
 
     custom_configuration = str()